Module Angstrom.Let_syntax

The Let_syntax module is intended to be used with the ppx_let pre-processor, and just contains copies of functions described elsewhere.

val return : 'a -> 'a t

val (>>|) : 'a t -> ('a -> 'b) -> 'b t

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

module Let_syntax : sig ... end